DSU: Add denoise worker, Bandit MPS fix, release workflow update
- Denoise worker: envelope-matched spectral subtraction, JSON protocol
- utils/denoise.py: main implementation (bundled in frozen build)
- noise-profile.wav included for dsu-denoise
- Bandit: fix float64 buffers for MPS (standard_window .float())
- build_dsu.py: numpy._core.tests for frozen denoise, keep numpy._core.tests in post-build
- Release: Mac ARM uses build_runtime_mac_mps.sh (mirror local), smoke test dsu-denoise

+++ models/bandit/core/utils/audio.py
@@ -392,7 +392,7 @@ class LinearFader(BaseFader):
         # using nn.Parameters allows lightning to take care of devices for us
         self.register_buffer(
                 "standard_window",
-                torch.concat([in_fade, center_ones, out_fade])
+                torch.concat([in_fade, center_ones, out_fade]).float()
         )
 
         self.fade_edge_frames = fade_edge_frames
@@ -438,9 +438,10 @@ class OverlapAddFader(BaseFader):
         )
 
         self.register_buffer(
-            "standard_window", torch.windows.__dict__[window_type](
-                    self.chunk_size, sym=False,  # dtype=torch.float64
-            ) / self.hop_multiplier
+            "standard_window", (
+                torch.windows.__dict__[window_type](self.chunk_size, sym=False)
+                / self.hop_multiplier
+            ).float()
         )
